# Support .NET Standard 2.0 and move to VS 2017