--- a/build.ps1
+++ b/build.ps1
@@ -2,7 +2,10 @@
 # creates NuGet package at \artifacts
 dotnet --version
 
-dotnet restore .\src\NLog\
-dotnet pack .\src\NLog\  --configuration release -o artifacts
+# dotnet restore .\src\NLog\
+msbuild /t:Restore /p:targetFrameworks='"net45;net40;net35;netstandard2.0;sl4;sl5;wp8;monoandroid44;xamarinios10"' .\src\NLog\
+
+# dotnet pack .\src\NLog\  --configuration release --include-symbols -o ..\..\artifacts
+msbuild /t:Pack /p:targetFrameworks='"net45;net40;net35;netstandard2.0;sl4;sl5;wp8;monoandroid44;xamarinios10"' /p:Configuration=Release /p:IncludeSymbols=true .\src\NLog\ /p:PackageOutputPath=..\..\artifacts
 
 exit $LASTEXITCODE