--- a/run-tests.ps1
+++ b/run-tests.ps1
@@ -1,0 +1,9 @@
+dotnet restore .\src\NLogAutoLoadExtension\
+dotnet restore .\tests\SampleExtensions\
+dotnet restore .\tests\NLog.UnitTests\
+dotnet build .\src\NLogAutoLoadExtension\ --configuration release 
+dotnet build .\tests\SampleExtensions\ --configuration release 
+dotnet build .\tests\NLog.UnitTests\ --configuration release 
+
+dotnet test .\tests\NLog.UnitTests\  --configuration release 
+exit $LASTEXITCODE