--- a/run-tests.ps1
+++ b/run-tests.ps1
@@ -1,11 +1,47 @@
+dotnet restore .\src\NLog\
+if (-Not $LastExitCode -eq 0)
+	{ exit $LastExitCode }
+
 dotnet restore .\src\NLogAutoLoadExtension\
+if (-Not $LastExitCode -eq 0)
+	{ exit $LastExitCode }
+
 dotnet restore .\tests\SampleExtensions\
-dotnet restore .\tests\NLog.UnitTests\
+if (-Not $LastExitCode -eq 0)
+	{ exit $LastExitCode }
 
-dotnet build .\src\NLogAutoLoadExtension\ --configuration release 
-dotnet build .\tests\SampleExtensions\ --configuration release 
-dotnet build .\tests\NLog.UnitTests\ --configuration debug
-dotnet build .\tests\NLog.UnitTests\ --configuration release 
+msbuild /t:Build /p:Configuration=Release .\src\NLogAutoLoadExtension\ /verbosity:minimal
+if (-Not $LastExitCode -eq 0)
+	{ exit $LastExitCode }
 
-dotnet test .\tests\NLog.UnitTests\  --configuration release 
-exit $LASTEXITCODE
+msbuild /t:Build /p:Configuration=Release .\tests\SampleExtensions\ /verbosity:minimal
+if (-Not $LastExitCode -eq 0)
+	{ exit $LastExitCode }
+
+msbuild /t:Restore,Build /p:Configuration=Release .\tests\NLog.UnitTests\ /verbosity:minimal
+if (-Not $LastExitCode -eq 0)
+	{ exit $LastExitCode }
+
+dotnet test .\tests\NLog.UnitTests\  --configuration release --framework netcoreapp2.0 --no-build
+if (-Not $LastExitCode -eq 0)
+	{ exit $LastExitCode }
+
+msbuild /t:Build /p:targetFramework=net452 .\tests\NLog.UnitTests\ /p:Configuration=Release /p:TestTargetFramework=net35 /verbosity:minimal
+if (-Not $LastExitCode -eq 0)
+	{ exit $LastExitCode }
+
+& ${env:xunit20}\xunit.console.x86.exe .\tests\NLog.UnitTests\bin\release\net452\NLog.UnitTests.dll -appveyor -noshadow
+if (-Not $LastExitCode -eq 0)
+	{ exit $LastExitCode }
+
+msbuild /t:Build /p:targetFramework=net452 .\tests\NLog.UnitTests\ /p:Configuration=Release /p:TestTargetFramework=net40-client /verbosity:minimal
+if (-Not $LastExitCode -eq 0)
+	{ exit $LastExitCode }
+
+& ${env:xunit20}\xunit.console.x86.exe .\tests\NLog.UnitTests\bin\release\net452\NLog.UnitTests.dll -appveyor -noshadow
+if (-Not $LastExitCode -eq 0)
+	{ exit $LastExitCode }
+
+msbuild /t:Build /p:Configuration=Debug .\tests\NLog.UnitTests\ /verbosity:minimal
+if (-Not $LastExitCode -eq 0)
+	{ exit $LastExitCode }