--- a/run-tests.ps1
+++ b/run-tests.ps1
@@ -26,19 +26,19 @@ dotnet test .\tests\NLog.UnitTests\  --configuration release --framework netcore
 if (-Not $LastExitCode -eq 0)
 	{ exit $LastExitCode }
 
-msbuild /t:Build /p:targetFramework=net452 .\tests\NLog.UnitTests\ /p:Configuration=Release /p:TestTargetFramework=net35 /verbosity:minimal
+msbuild /t:Build /p:targetFramework=net452 .\tests\NLog.UnitTests\ /p:Configuration=Release /p:TestTargetFramework=net35 /p:OutputPath=.\bin\release\net35 /verbosity:minimal
 if (-Not $LastExitCode -eq 0)
 	{ exit $LastExitCode }
 
-& ${env:xunit20}\xunit.console.x86.exe .\tests\NLog.UnitTests\bin\release\net452\NLog.UnitTests.dll -appveyor -noshadow
+& ${env:xunit20}\xunit.console.x86.exe .\tests\NLog.UnitTests\bin\release\net35\NLog.UnitTests.dll -appveyor -noshadow
 if (-Not $LastExitCode -eq 0)
 	{ exit $LastExitCode }
 
-msbuild /t:Build /p:targetFramework=net452 .\tests\NLog.UnitTests\ /p:Configuration=Release /p:TestTargetFramework=net40-client /verbosity:minimal
+msbuild /t:Build /p:targetFramework=net452 .\tests\NLog.UnitTests\ /p:Configuration=Release /p:TestTargetFramework=net40-client /p:OutputPath=.\bin\release\net40 /verbosity:minimal
 if (-Not $LastExitCode -eq 0)
 	{ exit $LastExitCode }
 
-& ${env:xunit20}\xunit.console.x86.exe .\tests\NLog.UnitTests\bin\release\net452\NLog.UnitTests.dll -appveyor -noshadow
+& ${env:xunit20}\xunit.console.x86.exe .\tests\NLog.UnitTests\bin\release\net40\NLog.UnitTests.dll -appveyor -noshadow
 if (-Not $LastExitCode -eq 0)
 	{ exit $LastExitCode }