--- a/src/LocalTestRun.vs2008.testrunconfig
+++ b/src/LocalTestRun.vs2008.testrunconfig
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<TestRunConfiguration name="Local Test Run" id="d5ad33a1-e178-4da8-880b-e73a8eedbdbe" xmlns="http://microsoft.com/schemas/VisualStudio/TeamTest/2010">
-  <Description>This is a default test run configuration for a local test run.</Description>
-  <TestTypeSpecific />
-</TestRunConfiguration>