--- a/src/NLog.Extended/NLog.Extended.csproj
+++ b/src/NLog.Extended/NLog.Extended.csproj
@@ -5,25 +5,17 @@
 
     <TargetFrameworks Condition=" '$(TargetFrameworks)' == '' ">net45;net40-client;net35</TargetFrameworks>
 
-    <VersionPrefix>4.5.0</VersionPrefix>
-    <VersionSuffix>beta01</VersionSuffix>
-    <AssemblyVersion Condition=" '$(AssemblyVersion)' == '' AND '$(APPVEYOR_BUILD_NUMBER)' != '' ">4.0.0.0</AssemblyVersion>
-    <AssemblyVersion Condition=" '$(AssemblyVersion)' == '' ">0.0.0.0</AssemblyVersion>
-    <APPVEYOR_BUILD_NUMBER Condition="'$(APPVEYOR_BUILD_NUMBER)' == ''">1</APPVEYOR_BUILD_NUMBER>
-    <FileVersion>$(VersionPrefix).$(APPVEYOR_BUILD_NUMBER)</FileVersion>
-    <InformationalVersion>$(VersionPrefix)</InformationalVersion>
-    <InformationalVersion Condition="'$(VersionSuffix)' != ''">$(InformationalVersion)-$(VersionSuffix)</InformationalVersion>
-
     <Title>NLog for unknown platform Extended Profile</Title>
     <Company>NLog</Company>
     <Description>NLog</Description>
-    <Product>NLog Extended Profile v$(VersionPrefix)</Product>
-    <Product Condition="'$(VersionSuffix)' != ''">$(Product)-$(VersionSuffix)</Product>
+    <Product>NLog Extended Profile v$(ProductVersion)</Product>
+    <InformationalVersion>$(ProductVersion)</InformationalVersion>
     <Authors>Jarek Kowalski,Kim Christensen,Julian Verdurmen</Authors>
     <CurrentYear>$([System.DateTime]::Now.ToString(yyyy))</CurrentYear>
     <Copyright>Copyright (c) 2004-$(CurrentYear) NLog Project - http://nlog-project.org/ </Copyright>
-    
+
     <SignAssembly>true</SignAssembly>
+    <AssemblyVersion>4.0.0.0</AssemblyVersion>
     <AssemblyOriginatorKeyFile>..\NLog.snk</AssemblyOriginatorKeyFile>
     <PublicSign Condition=" '$(OS)' != 'Windows_NT' ">true</PublicSign>