--- a/src/NLog.Extended/Properties/AssemblyInfo.cs
+++ b/src/NLog.Extended/Properties/AssemblyInfo.cs
@@ -36,24 +36,6 @@ using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-#if NET3_5
-    [assembly: AssemblyTitle("NLog for .NET Framework 3.5 Extended Profile")]
-#elif NET4_0
-    [assembly: AssemblyTitle("NLog for .NET Framework 4 Extended Profile")]
-#elif NET4_5
-    [assembly: AssemblyTitle("NLog for .NET Framework 4.5 Extended Profile")]
-#elif MONO_2_0
-    [assembly: AssemblyTitle("NLog for Mono 2.0 Extended Profile")]
-#elif DOCUMENTATION
-    [assembly: AssemblyTitle("NLog Documentation")]
-#else
-#error Unrecognized build target - please update AssemblyInfo.cs
-#endif
-
-[assembly: AssemblyDescription("NLog")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("NLog")]
-[assembly: AssemblyCopyright("Copyright (c) 2004-2016 Jaroslaw Kowalski, Kim Christensen, Julian Verdurmen")]
 [assembly: AssemblyCulture("")]
 
 [assembly: CLSCompliant(true)]