… .travis.yml
… appveyor.yml
… src/LocalTestRun.netfx40.testrunconfig
… src/LocalTestRun.netfx45.testrunconfig
… src/Mono.targets
… src/NLog.Extended/NLog.Extended.doc.csproj
… src/NLog.Extended/NLog.Extended.mono.csproj
… src/NLog.Extended/NLog.Extended.netfx20.csproj
… src/NLog.Extended/NLog.Extended.netfx35.csproj
… src/NLog.Extended/NLog.Extended.netfx40.csproj
… src/NLog.Extended/NLog.Extended.netfx45.csproj
… src/NLog.Extended/ProjectFileInfo.xml
… src/NLog.Xamarin.Android.sln
… src/NLog.Xamarin.iOS.sln
… src/NLog.doc.sln
… src/NLog.mono.sln
… src/NLog.netfx35.sln
… src/NLog.netfx40.sln
… src/NLog.netfx45.sln
… src/NLog.proj
… src/NLog.sl4.sln
… src/NLog.sl5.sln
… src/NLog.wp8.sln
… src/NLog/Common/AsyncHelpers.cs
… src/NLog/Common/InternalLogger-generated.tt
… src/NLog/Common/InternalLogger.cs
… src/NLog/ILogger.cs
… src/NLog/ISuppress.cs
… src/NLog/Internal/ConfigurationManager.cs
… src/NLog/Internal/Fakeables/AppDomainWrapper.cs