Initial attempt for converting to new VS2017 csproj file format.

- [x] NetStandard 2.0
- [x] Net45, Net40-client, Net35 (Moved net40 to net40-client nuget-lib-folder)
- [x] Silverlight 4, Silverlight 5, Windows Phone
- [x] XamarinIOS 1.0
- [x] MonoAndroid (Upgraded from 2.3 to 4.4)
- [x] OpenCover
- [x] Travis NetStandard 2.0
- [x] Travis Mono 5.2 Net45 (Ver. 5.2 needed for msbuild support)
- [x] Cherry picking from coreclr-branch (NLog 5.0 BETA)
- [x] Strong Name Version remains static (ver. 4.0.0.0)
- [x] xUnit 2.3.0-beta5 (BETA Needed for NetStandard 2.0)
- [x] xUnit 2.0 testing of net35 and net40-client
- [x] ~NLog-Docs-csproj (Maybe handled by GenerateDocumentationFile=true)~ -> not used
- [x] ~NLog-BuildDocPages~ -> not used for now
- [x] NLog-CheckSourceCode
- [x] NLog.Extended package
- [x] NLog.Schema package (XSD Generation for config-file-intellisense) -> Julian
- [x] NLog.Config package
- [x] NLog.Snippets package (Maybe it is replaced by https://github.com/NLog/NLogSnippetsVSIX)
- [x] NLog.shfbproj (Sandcastle help file) -> revert

Replaces #2175 and #2154 and resolves #2152, resolves #1954 and resolves  #1953