@luigiberrettini 2017-08-24
Just to avoid confusion I would rename this PR as "VS 2017 and .NET Standard 2.0"