@snakefoot 2017-08-26
@luigiberrettini You can now play around with this nuget-package:

https://ci.appveyor.com/project/nlog/nlog/build/4.5.0-beta5831/artifacts