@snakefoot 2017-08-26
@304NotModified It would be nice if this PR is merged before any other PR, as the conflict rate is quite high with this one.