@snakefoot 2017-08-26
@304NotModified Seems that the AppVeyor-build-image for VS2017 includes NetCore2, but VS2017 has dropped support for Silverlight. And the AppVeyor-build-image for VS2015 does not include NetCore 2.0 SDK.

Any ideas how to solve this? Moving forward without Silverlight ? Am I missing a magic parameter?