@snakefoot 2017-08-27
@304NotModified Well after solving the issue with Silverlight and WindowsPhone, then I'm now stuck at MonoDroid and MonoTouch. Don't know how to make them build. But I guess we can live without them while doing the NLog 4.5 BETA exercise?