@snakefoot 2017-09-10
Made a very crude comparison of build-timings:

Task | master | This PR
----|---------|--------
NLog.Schema.nupkg | 110 sec | 100 sec
UnitTest - NetStandard |  | 100 sec (2060 tests)
UnitTest - Net35 | 130 sec (2154 tests) | 179 sec (2205 tests)
UnitTest - Net40 | 130 sec (2192 tests) | 177 sec (2215 tests)
UnitTest - Net45 | 311 sec (2223 tests) | 415 sec (2223 tests)